TIP-856: Move Pim\Buffer classes resources to Akeneo\Tool
