+++ src/Pim/Component/Connector/Writer/File/Xlsx/Writer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Component\Connector\Writer\File\Xlsx;
 
-use Akeneo\Component\Buffer\BufferFactory;
+use Akeneo\Tool\Component\Buffer\BufferFactory;
 use Akeneo\Tool\Component\Batch\Item\FlushableInterface;
 use Akeneo\Tool\Component\Batch\Item\InitializableInterface;
 use Akeneo\Tool\Component\Batch\Item\ItemWriterInterface;
